requirements.txt
asgiref==3.2.3
boto3==1.10.32
botocore==1.13.32
coverage==4.5.4
dj-database-url==0.5.0
Django==3.0
django-heroku==0.3.1
django-nose==1.4.6
django-storages==1.8
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
jmespath==0.9.4
nose==1.3.7
Pillow==6.2.1
psycopg2==2.7.7
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
